SV-003 click on hamburger menu
